Extract unchanged upload helpers to satisfy SDK size ratchet
